requirements.txt
fastapi[all]
streamlit==1.29.0
pandas==2.1.4
python-dotenv==1.0.0
openai==1.6.1
langchain==0.0.354
langchain-experimental==0.0.47
tabulate==0.9.0